requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchain-text-splitters
pypdf
chromadb
langchain-openai